Increase the wait time to process the keyword

+++ priv/repo/migrations/20200414103012_create_links_and_add_links_count_to_keywords.exs
@@ -3,7 +3,7 @@ defmodule GoogleCrawler.Repo.Migrations.CreateLinksAndAddLinksCountToKeyword do
 
   def change do
     alter table(:keywords) do
-      add :total_results, :integer
+      add :total_results, :bigint
       add :total_ads_links, :integer
       add :total_links, :integer
     end
